backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic-settings==2.9.1
PyJWT==2.10.1
python-multipart==0.0.20
openai==1.76.2
email-validator==2.2.0
